requirements.txt
pandas
requests
python-dotenv
markdown
pydub 